requirements.txt
openai>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pyyaml>=6.0.0
selenium>=4.0.0
